+++ docs/docs/changelog.md
@@ -1,15 +1,22 @@
 ## 更新日志
 
-### 2.2.1
+### 2.3.0
 
 #### 新特性
 
 - CalendarView
   - 新增日历面板组件 (by [@yawuling](https://github.com/yawuling) ) )
+- ColPicker
+  - 新增 `z-index` 属性，设置弹层层级 (by [@yawuling](https://github.com/yawuling) ) )
 - DatetimePicker
   - 新增 `default-value` 属性，支持设置面板默认展示值 (by [@yawuling](https://github.com/yawuling) ) )
+  - 新增 `z-index` 属性，设置弹层层级 (by [@yawuling](https://github.com/yawuling) ) )
 - Input
   - 添加自定义外部类名 `custom-textarea-container-class` (by [@yawuling](https://github.com/yawuling) ) )
+- Picker
+  - 新增 `z-index` 属性，设置弹层层级 (by [@yawuling](https://github.com/yawuling) ) )
+- SelectPicker
+  - 新增 `z-index` 属性，设置弹层层级 (by [@yawuling](https://github.com/yawuling) ) )
 - Switch
   - 新增 `before-change` 修改前钩子 (by [@yawuling](https://github.com/yawuling) ) )
 - Toast
@@ -25,6 +32,8 @@
   - 优化选中项样式 (by [@yawuling](https://github.com/yawuling) ) )
 - StatusTip
   - 占位图改用网络地址，小程序打包时会将未使用的组件也打包进去，如果使用本地图片，会明显增大小程序包体积 (by [@yawuling](https://github.com/yawuling) ) )
+- 定位层级
+  - 优化多个组件的定位层级 (by [@yawuling](https://github.com/yawuling) ) )
 
 #### Bug 修复
 
@@ -38,6 +47,7 @@
   - 修复自定义 `label` 插槽失败的问题 (by [@yawuling](https://github.com/yawuling) ) )
 - DatetimePicker
   - 修复 `error` 错误状态时 placeholder 未标红问题 (by [@yawuling](https://github.com/yawuling) ) )
+  - 修复 `open` 和 `close` 无法调用问题 (by [@yawuling](https://github.com/yawuling) ) )
 - MessageBox
   - 修复 `closeOnClickModal` 无效问题 (by [@yawuling](https://github.com/yawuling) ) )
   - 修复 `messageBox` 使用 `type` 无效问题 (by [@yawuling](https://github.com/yawuling) ) )